Fix: getGoogleUser함수 lib으로 옮김
