+++ backend/src/services/auth.sevice.ts
@@ -1,16 +1,5 @@
 import { prisma } from '../plugins/prisma';
 
-export async function getGoogleUser(accessToken: string) {
-  const res = await fetch('https://www.googleapis.com/oauth2/v2/userinfo', {
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-    },
-  });
-
-  if (!res.ok) throw new Error('Google user fetch failed');
-  return await res.json();
-}
-
 export async function handleGoogleUser(googleUser: any) {
   const { email, name, picture } = googleUser;
